+++ packages/video_player/video_player/lib/video_player.dart
@@ -469,6 +469,12 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
       );
     }
 
+    if (videoPlayerOptions?.allowBackgroundPlayback != null) {
+      await _videoPlayerPlatform.setAllowBackgroundPlayback(
+        videoPlayerOptions!.allowBackgroundPlayback,
+      );
+    }
+
     _playerId =
         (await _videoPlayerPlatform.createWithOptions(creationOptions)) ??
         kUninitializedPlayerId;