+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayerPlugin.java
@@ -188,6 +188,11 @@ public class VideoPlayerPlugin implements FlutterPlugin, AndroidVideoPlayerApi {
         : flutterState.keyForAssetAndPackageName.get(asset, packageName);
   }
 
+  @Override
+  public void setAllowBackgroundPlayback(boolean allowBackgroundPlayback) {
+    sharedOptions.allowBackgroundPlayback = allowBackgroundPlayback;
+  }
+
   private interface KeyForAssetFn {
     String get(String asset);
   }