+++ packages/video_player/video_player_android/lib/src/android_video_player.dart
@@ -225,6 +225,11 @@ class AndroidVideoPlayer extends VideoPlayerPlatform {
     return _api.setMixWithOthers(mixWithOthers);
   }
 
+  @override
+  Future<void> setAllowBackgroundPlayback(bool allowBackgroundPlayback) {
+    return _api.setAllowBackgroundPlayback(allowBackgroundPlayback);
+  }
+
   _PlayerInstance _playerWith({required int id}) {
     final _PlayerInstance? player = _players[id];
     return player ?? (throw StateError('No active player with ID $id.'));